tutorial: "Make a commit"
Tutorial narrative is poor.  Says "now make some changes!"  This is followed by a subtitle that says "To commit changes"

So I'm trying to figure out how to make those changes so I can commit them.  I knew I saw an EDIT button eralier somewhere, so I'm looking for the EDIT button.  I see lots of other buttons, but no EDIT.  I finally figure out that its in egyptian hyrogliphics  as a little picture of a pencil.pen/crayon/whatever.  I click it and make the changes.

Now I read on to see how "To commit changes".  It says:

To commit changes

    Click the README file.
    Click the pencil icon in the upper right corner of the file view to edit.
    In the editor, write some text, tell a bit about yourself.
    Write a commit message that describes your changes.

So what is titles "To commit changes" is really how to make change and commit them.  I am beginning to get the impression that the guy who wrote this tutorial was an "English as a second language" guy.  :-(

Enlarging the screenshot, I see the guy was from another planet.  That explains the poor english expository prose.  :-(